[OSCAR::Utils] Only attempt to kill if some processes have been found holding a lock on selected path.
git-svn-id: http://svn.oscar.openclustergroup.org/oscar/trunk@10981 18f9d06c-6ae2-0310-8438-938a554feff9

Former-commit-id: 99c31c6d895e90705970a59c35f753df348844d3 [formerly d689d1e90d9d176fb2641245b619d3cd94b8d5a9] [formerly 0735b04acdc32428234d371a7b1926159669b94b [formerly 3cdd29ca4fa868132f5d6246cc855df9a8ea0847]] [formerly fdfe6cf580dfce012cb95ec704a51a5df300c6bd [formerly ce8fe31f41dc5c4f8f10fa4486d7d9dfd24e3032] [formerly db9ac752dcbfb5e0129ebfe613b67cc7f4621ec6 [formerly 23009798d2914defa287b58b43d95916cf9193a3]]]
Former-commit-id: 8e58797f015bbb00307ab4bf7b7059b55ae296e2 [formerly 6de89268f088db1ca82e965a2df1dd97e249885f] [formerly edb6ea360884641948dd5883a889d501a6946e1c [formerly daea099ef7db19c2833a5b129d41a24a03a12ac5]]
Former-commit-id: 772f5b5dfcea8b851b19a2d1b0d04235e8fb4630 [formerly c14d24e4ce0336a0dd7b3956f28ae3a0e43f5534]
Former-commit-id: 971de8dd5885bd2bfe5529e90af5bc419cfa5368

+++ lib/OSCAR/Utils.pm
@@ -112,11 +112,15 @@ sub kill_processes_locking_path($) {
     }
     close LSOF;
 
-    oscar_log(5, INFO, "Sending 'SIGTERM' to " . join(", ",@to_kill));
-    kill 'TERM',@to_kill;
+    if(@to_kill) {
+        oscar_log(5, INFO, "Sending 'SIGTERM' to " . join(", ",@to_kill));
+        kill 'TERM',@to_kill;
 
-    oscar_log(5, INFO, "Sending 'SIGHUP' to " . join(", ",@to_kill));
-    kill 'HUP',@to_kill; # Shell scripts will ignore TERM.
+        oscar_log(5, INFO, "Sending 'SIGHUP' to " . join(", ",@to_kill));
+        kill 'HUP',@to_kill; # Shell scripts will ignore TERM.
+    } else {
+        oscar_log(5, INFO, "No process to kill.");
+    }
 }
 
 ################################################################################
